init: create base settings of the project
